Changed edit method
